Need gamma_min ~ 0.1 for grtrans and symphony to agree.
Not quite sure why this is, but I don't think it affects the performance of
the calculations at all so whatever. Once this change is made, agreement
between the grtrans approximations and symphony calculations is quite good! I
believe the grtrans polynomials are derived directly from the symphony results
so that's not shocking, but nice to confirm.