pylib/geometry.py: allow non-uniform electron densities in SimpleWasherDistribution
In particular, we allow n_e to vary as a power law in r such that it increases
towards the inner edge of the annulus. We do some math to make it so that
changing the concentration parameter does not affect to the total number of
electrons in the washer.