pylib: start working on proper linear Stokes rotation
I've needed to do this for a while. I haven't actually implemented the key
rotation step, but I think the framework is in place to calculate the needed
angles and propagate them where they need to go. The only testing thus far
has been me writing things on paper, though.