pylib: finish implementing correct linear polarization, maybe
I've implemented a function that attempts to rotate the radiative transfer
coefficients correctly to get linear polarization right. Trouble is, I have
absolutely no idea if I've done it right. And I have no idea how to test it. I
guess that's OK if the total amplitude of the linpol is small ...