pylib/vkl.py: add a full 3D solution
I don't expect to be able to solve the problem in realistically in full 3D,
but we can use a small grid to at least get *something* going.

A lot of code duplication here; should probably tidy that up since I think we
will still want to run 3D solutions even when (2+1)D computations work --
they'll be a useful double-check.