Solve for a particle distribution!
The solution algorithm doesn't actually care about the Numerical object at
all, so it makes sense to put it on the ThreeDCoordinates.

I have multiplied parameters by all sorts of crazy values to get the solution
to look kinda sorta vaguely realistic. One day we will have to go back and
sort through all of that, but this is good enough to get us to the point where
we glue together the particle simulation and the radiative transfer.