Export V/K/L-space distributions to a format aimed at the RT calculator
I think I'm taking the right tack here. Given a VKL distribution, we convert
it into a *4*-dimensional distribution in L, magnetic latitude, energy, and
pitch angle. For each VKL cube cell we smear out its contribution along a line
in the latitude/pitch-angle plane, which represents how those quantities
evolve as it bounces around. The RT code meanwhile will be tracing rays down
the L/latitude axis. For each location, there will be an observed distribution
of energies and pitch angles that we will hopefully be able to express in a
way that lets us use Symphony to determine the RT coefficients.