pylib/vkl.py: almost complete restructure
The prior version of this module tied the coefficient computations much too
closely to dolfin. This version tries to make it so that you can compute all
of the key values given a basic Coordinates object whose details are not
tightly specified.

Also, I think I was mis-defining K by including an R_Earth term, which
explains why my Cg is so different than the one in Subbotin & Shprits 2012.
Needs more investigation since I was looking at its definition pretty closely
before.

pylib/dolfin.py should be transitioned to build on this module if I keep on
using it. Unclear whether that will be the case, though.