pylib/ljob.py: use Python >= 3.4's `selectors` module
I ran into a problem since standard Unix select(2) can apparently only handle
FD numbers less than 1024, and I was exceeding this with a big job! The
`selectors` module genericizes the functionality and uses epoll() on Linux,
which does not have this limitation.

However I've been lazy and not switched the worker code. Workers are just
going to have one socket open, not a thousand.