pylib/ljob.py: be more stringent with workers than with tasks
If there are similar numbers of tasks and workers, you can get a corner case
if there's a worker that always fails at its tasks. The worker can fail at
task N, then immediately get re-assigned the same task, fail at it again, and
so on. The way I had things structure, the manager would declare the task
un-completable before it would decide that the worker was busted, and so you
would get tasks erroneously marked as failed.

Here I change the thresholds such that the worker will be marked failed before
the task is. That way, the worker will be rescheduled and it is much more
likely that the task will be reattempted on a different, hopefully not busted,
worker. This doesn't eliminate the possibility of tasks being marked as
failures when they really can be completed (pretty sure that's not possible),
but it should help.