sde/grid.py: rework parameters fed into Summers/Shprits calculation
Many of the parameters were things that scaled with the cyclotron frequency,
and so we were having waves that got much stronger as you got closer to the
body, and so on. I don't think that's physically very plausible, so now we
feed some made-up physical choices that are independent of L. Precise values
will quite likely demand tweaking once I get a clearer picture of how the
simulations equilibrate.