Take the plunge and switch to the g/alpha/L coordinate system.
I *think* this will be good because I've been a bit worried about how my
derivatives and such interact with a non-uniform sampling in the former "p"
coordinate. We write `g = ln(p / m_e c)` and transform coordinates everywhere
as appropriate.

We also do the work to figure out the full Jacobian of the MJΦ → gαL
coordinate change, so that we can properly add tiny diagonal terms to the
diffusion matrix representing the radial diffusion effect. Without these, it's
not really a diffusion matrix since those have to be positive definite.