pylib/geometry.py: add a simple mechanism for oversampling rays near the body
The hope is that this addition to the SimpleRayTracer will avoid the NaNs I've
been getting. I tried to use the FormalRayTracer to avoid this since it has a
much better-motivated algorithm, but I forgot that it needs to evaluate the
actual synchrotron coefficients that will be fed into the radiative transfer
calculation. That's possible but the preprays/integrate separation aims to
avoid those calculations at this stage.

So, instead we add a very naive hack that just amps up the sampling resolution
near the body. Time will tell whether this actually helps. If it doesn't, it
won't be the end of the world to extend preprays to know about the synchrotron
coefficients and the frequencies that are being modeled. Hopefully the
additional computational cost wouldn't be so bad either ...