Make the preprays HDF5 storage format much more efficient
When I added higher resolution to avoid NaNs in the center the output file
ballooned up to 190 GB. Disk space is cheap but that's a bit ridiculous.

The penalty is that the "assemble" step now has to visit all of the numpy
intermediate files twice, which makes it a lot slower. I think it'll be
worth it in the end.