Work on modeling with KZN-based particle distributions
The first part of this is adapting to the new output format of the KZN code.
The second part is that this spurred an attempt to start rationalizing how the
configuration system works. There are elements of its design that I definitely
forget why exactly I did things that way.