vernon/integrate.py: wire up a cold plasma contribution
This is a bit of a hack, since the cold plasma properties are set in the
`[rt]` section of the configuration isn't really appropriate. Also I doubt
that a strictly proportional, spatially invariant relationship between cold
and hot electrons is the most realistic model. But this will get us going.