vernon/kzn.py: implement self-consistent density functions
I thought that the eta2 parameter in the KZN model was unspecified, which
meant that you could choose any particle distribution that you wanted. But
really it is implicitly defined by the choice of \chi and \Omega_*, as
suggested by Eq. 21 in the paper. Here we implement the analysis to calculate
\eta_2, and therefore the particle density distribution, from the loaded model
and the relevant physical parameters.

Right now we have a gross hack in that the KZNKWKDistribution computes the
distribution of cold electrons, then decimates it by a fixed factor to get the
density of hot electrons. The radiative transfer code then takes the hot
electron density and scales it back up to get a cold electron density. The
basic problem is that the code is not wired up to ship around different cold
and hot electron densities. Shouldn't be too hard to fix but I didn't want to
deal with that issue right now.